Rename ReconstructLeafHash to ToEntryHash and merge into hashedrekord.go
Move the function from its own file into hashedrekord.go alongside ToLogEntry,
trim the doc comment, and consolidate tests into hashedrekord_test.go. Delete
the now-empty leafhash.go and leafhash_test.go files.
